include/TermNode: Fixed nested call safety in term container clearing.
The naive implementation using recursion calls in C++ would cause undefined behavior (native stack overflow typically) when the term node has a large amount of nested subnode levels. Since the term node is also a representation of an object in the object language which can be constructed in the user program, this is a security hole.
The fix uses a loop to transfer subnodes by linearization, which flatten the tree structure. Thus, no native calls depend on the depth of the nested levels, and there is a limitation of the depth of nested active calls.

Signed-off-by: He Bin <hebin@uniontech.com>