src/Evaluation.cpp: Avoided trailing bindings to have the temporary term tag transferred from the unique term tag.
A term represents a temporary object when it has been bound with '&' or '%' sigil by adding the temporary tags for prvalue operands. The tags is also observable via references to such objects, to trace the source of the operand. In particular, reference collapsing in the variable binding (binding an object having reference type from a reference value operand using the '&' or '%' sigil) operates on the tags transferred from the operand. In the case of using with '&' sigil of reference collapsing, an xvalue (having the unique tag in the reference operand) is additionally converted to make the bound operand having the temporary tag, and the unique tag will be invisible after the name resolution (i.e. all named variable used directly are lvalues). Thus, all rvalues (including xvalues and prvalues) are treated the same in this sense.
The temporary tag saved in the reference value essentially mimics the additional special type information in the forwarding references (rvalue references in a form of 'T&&' where 'T' is a template type parameter) deduced from rvalue operands in C++ perfect forwarding, without the need of additional channel to encode this special usage. Such usage allows the object language function 'forward!' (functioning like 'std::forward<decltype(T)>(some_operand)' in C++) implementable as an applicative, rather than an operative (named like '$forward!' by convention) using the contextual information provided by the value category (i.e. the tags) of the operand to be forwarded.
The temporary tag is essentially of the type information of value category (specifically, prvalues) transferred in the per-object runtime metadata. In the cases other than being the operand of 'forward!'-like operations, it is often unexpected. In such cases, the unique reference (xvalues, also working like xvalues in C++) should be used instead, to avoid unexpected moving from lvalue operands (which might have the temporary tag from elsewhere). In other words, temporary tag on references (not direct on objects implying prvalues) should often be removed in time. This is done in cases like '&' binding for a non-reference list suboperand.
In the cases of binding a trailing operand, each binding is delegated to its suboperands, and the condition would be inspected on the each suboperand to determine which tags the bound suboperand should have. There is no way to remove the temporary tag as non-trailing operands binding with '&' sigil. As a result, it is practically unexpected to have such special forwarding rule, because there is also no alternative way to operate on subobject of a bound list (as perfect forwarding in C++). Moreover, there is no other similar feature in C++. Removing the temporary tag by other means in the object language would cause some additional headache (if needed), so keep the tag conversion away directly in the binding rule is a better choice.

Signed-off-by: He Bin <hebin@uniontech.com>