include/Interpreter.h: Enabled memory pooling via the polymorphic pool memory resource.
The time performance of the result interpreter in the release configurations should be better in most cases, due to less needs of the allocation and deallocation operations. However, it can be slow for debug configurations because the pool resource implementation from
YBase is also not built with the compiler optimization options enabled.

Signed-off-by: He Bin <hebin@uniontech.com>