src/Main.cpp: Simplified the main function.
This also fixes the linker failure using either "build.sh" or debug configurations with "sbuild.sh".
The former does not work because there is missing source files.
The latter seems a tool chain bug (at least with GCC 8.3.0 + GNU ld 2.31.1) which makes the linker fails to ignore the static initialization of the indirectly referenced object "YSLib::IO::PreserveModificationAndAccessTime" (in the main function) and the function using to initialize it is actually exposed in the dynamic library being linked to.

Signed-off-by: He Bin <hebin@uniontech.com>