Fixed potential undefined behavior by missing the guarantee of the object destroying.
Failure to cleanup the continuations may cause the REPL misbehaves for subsequent inputs after an exception is handled.
This can also cause invalid object accesses in the remaining continuation frames (e.g. TCO actions) after the term destruction. For instance, let "X" and "X" be names of a non-existing module and an existing one, here is a case in REPL probably easy to reproduce in the previous versions of the interpreter:
$import! std.modules require;
require "X";
require "E";
exit
The program above might crash after 'exit' due to accessing invalid object referenced in TCOAction on destruction of the interpreter.

Signed-off-by: He Bin <hebin@uniontech.com>