Implemented the literal '#ignore' according to the new rules.
This allows it to be used as a variable name in contexts which differentiate the special placeholder with an identifier being defined (e.g. '$def!' and parameter passing). Moreover, is more consistent to other values being handled currently, hence simplifies some checks in these contexts.
The routine of sigil extraction is modified to allows empty symbol.

Signed-off-by: He Bin <hebin@uniontech.com>