include/TermNode.h: Propagated allocators on term container copy.
This is necessary to support the planned change to make the setting of term node container using swapping instead of assignment. Otherwise, the allocators would be easily mismatched and the narrow contract of allocator-aware container swapping (as the current design of ISO C++ standard allocator-aware containers) would be violated. Typically the program would crash once different allocators are used in such case.

Signed-off-by: He Bin <hebin@uniontech.com>