Lifted the indirect value for the <parent> parameter in functions.
The original implementation in the commit 45febec3966feddc290980e52034297d15445e7a did not support the single reference to environment occasionally. This was unfortunately inconsistent, because references to the environment list or references to the elements in the environment lists were supported.
The change is necessary to support more succinct implementation of the derivation of combiner constructors.
The imprecise descriptions of <eformals> and <parent> in the documentation (not even precise for the original implementations first introduced) are also fixed.

Signed-off-by: He Bin <hebin@uniontech.com>