src/TCO.cpp: Avoided to relied on the unspecified destruction order of 'forward_list::remove_if'.
This is not guaranteed for 'std' one in ISO C++, albeit known implementations just work.
This implementation also works for LWG 526 while older versions of libstdc++ not (see https://gcc.gnu.org/bugzilla/show_bug.cgi?id=91620). Although it is actually not required, but just keep as-is, and it seems even more efficient in the generated code.

Signed-off-by: He Bin <hebin@uniontech.com>