src/Evaluation.cpp: Simplified and optimized the internal implementation paths of sequence binding further.
This expands the internal implementation path for the binding without sigil and merges it with the forwarding paths, as well as the remained paths for binding references. These 2 merged group of paths are equivalent, but there are less branches for each case. Thus, the implementation is optimized.
Then, the resulted 2 groups of the paths are merged to extract common path of term creation.

Signed-off-by: He Bin <hebin@uniontech.com>