src/Forms.cpp: Refactored to adapt to the planned environment parent type changes with optimized parent list detection.
This is equivalent to the previous implementation in functionality. The case of single element in the parent list is now detected on construction and it is treated as a single element of parent instead of a parent list. It is also easier to change the parent type by using Unilang::ToParent which is designed to be more neutral to the target type in the environment parent source parameter.

Signed-off-by: He Bin <hebin@uniontech.com>