src/Main.cpp: Fixed missing preventing return value conversion on the forwarded body in the derivation of functions: $let, $let%.
This was a bug that would affect the body evaluation if the body can be a pair. This case would be ill-formed as per the language specification. Since the check of the codebase is not yet finished, the case to distinguish the effect of this fix is deliberately not provided.
For other operatives in the '$let' family, there exists the similar issue. This is not yet checked, either.
This was broken since commit fca1106f307ba5cd28043a4275b57bbfa9da1453, where it should have been changed as other functions.

Signed-off-by: He Bin <hebin@uniontech.com>