{include, src}/Context: Added 'const' on GlobalState::Read.
This is similar to GlobalState::ReadFrom. There is no need to mutate the internal global state. Unrestricted mutation will incur the GIL (global interpreter lock) after multithreading support is added (unless there is more fine-grained locking), so missing 'const' is potentionally bad. Anyway, one of the key point to introduce GlobalState is to reduce the need of GIL, exactly.

Signed-off-by: He Bin <hebin@uniontech.com>