install-sbuild.sh: Worked around Clang++ builds.
The updated configuration for running 'sbuild.sh' implies '-fuse-ld=lld' when the compiler is detected as Clang++. However, LLD in some platforms may be too buggy to work. So, disable it by patching during the setup of the dependencies.
This has no effects on 'build.sh' which does not tweak '-fuse-ld' at all, though.

Signed-off-by: He Bin <hebin@uniontech.com>